+++ src/main.rs
@@ -331,7 +331,12 @@ async fn run(args: Args) -> Result<()> {
             "macos" => "osx",
             other => other,
         };
-        let os_version = sysinfo::System::os_version().unwrap_or_else(|| String::from("0"));
+
+        let os_version = match std::env::consts::OS {
+            "linux" => sysinfo::System::kernel_version(),
+            _ => sysinfo::System::os_version(),
+        }
+        .unwrap_or_else(|| String::from("0"));
         if os_name.is_empty()
             || os_name.contains(illegal_chars)
             || os_version.is_empty()